Reduce testscript output

+++ testscript.sh
@@ -75,10 +75,14 @@ for f in $(ls $TESTDIR/$INPUTS); do
   p=0
   ./main.byte -i $EXTEND_FILE > $INTERPRETER_TARGET 2>&1
   ./main.byte -c $EXTEND_FILE > $EXTEND_TARGET 2>&1
-  llc-3.8 -filetype=obj $EXTEND_TARGET -o $COMPILED_OUTPUT
-  gcc -o tmp/tmp $COMPILED_OUTPUT tmp/std.o -lm
-  rm $COMPILED_OUTPUT
-  ./tmp/tmp > $TEXT_OUTPUT
+  llc-3.8 -filetype=obj $EXTEND_TARGET -o $COMPILED_OUTPUT > /dev/null 2>&1
+  if [ $? -eq 0 ]; then
+    gcc -o tmp/tmp $COMPILED_OUTPUT tmp/std.o -lm
+    rm $COMPILED_OUTPUT
+    ./tmp/tmp > $TEXT_OUTPUT
+  else
+    touch $TEXT_OUTPUT
+  fi
   diff $INTERPRETER_TARGET $EXPECTED_OUTPUT > $RESULT_OUTPUT 2>&1
   if [ $? -eq 0 ]; then
     counteri=$((counteri+1))
